+++ src/dataset.rs
@@ -8,15 +8,31 @@ use crate::{
 use gdal_sys::{
     self, CPLErr, GDALAccess, GDALDatasetH, GDALMajorObjectH, OGRErr, OGRLayerH, OGRwkbGeometryType,
 };
-use libc::{c_double, c_int};
+use libc::{c_double, c_int, c_uint};
 use ptr::null_mut;
 
 use crate::errors::*;
 use std::convert::TryInto;
 
+use bitflags::bitflags;
+
+/// A 2-D affine transform mapping pixel coordiates to world
+/// coordinates. See [GDALGetGeoTransform] for more details.
+///
+/// [GDALGetGeoTransform]: https://gdal.org/api/gdaldataset_cpp.html#classGDALDataset_1a5101119705f5fa2bc1344ab26f66fd1d
 pub type GeoTransform = [c_double; 6];
 static START: Once = Once::new();
 
+/// Wrapper around a [`GDALDataset`][GDALDataset] object.
+///
+/// Represents both a [vector dataset][vector-data-model]
+/// containing a collection of layers; and a [raster
+/// dataset][raster-data-model] containing a collection of
+/// rasterbands.
+///
+/// [vector-data-model]: https://gdal.org/user/vector_data_model.html
+/// [raster-data-model]: https://gdal.org/user/raster_data_model.html
+/// [GDALDataset]: https://gdal.org/api/gdaldataset_cpp.html#_CPPv411GDALDataset
 #[derive(Debug)]
 pub struct Dataset {
     c_dataset: GDALDatasetH,
@@ -28,6 +44,83 @@ pub fn _register_drivers() {
             gdal_sys::GDALAllRegister();
         });
     }
+}
+
+// These are skipped by bindgen and manually updated.
+#[cfg(major_ge_2)]
+bitflags! {
+    /// GDal extended open flags used by [`Dataset::open_ex`].
+    ///
+    /// Used in the `nOpenFlags` argument to [`GDALOpenEx`].
+    ///
+    /// Note that the `GDAL_OF_SHARED` option is removed
+    /// from the set of allowed option because it subverts
+    /// the [`Send`] implementation that allow passing the
+    /// dataset the another thread. See
+    /// https://github.com/georust/gdal/issues/154.
+    ///
+    /// [`GDALOpenEx`]: https://gdal.org/doxygen/gdal_8h.html#a9cb8585d0b3c16726b08e25bcc94274a
+    pub struct GdalOpenFlags: c_uint {
+        /// Open in read-only mode (default).
+        const GDAL_OF_READONLY = 0x00;
+        /// Open in update mode.
+        const GDAL_OF_UPDATE = 0x01;
+        /// Allow raster and vector drivers to be used.
+        const GDAL_OF_ALL = 0x00;
+        /// Allow raster drivers to be used.
+        const GDAL_OF_RASTER = 0x02;
+        /// Allow vector drivers to be used.
+        const GDAL_OF_VECTOR = 0x04;
+        /// Allow gnm drivers to be used.
+        #[cfg(any( all(major_ge_2,minor_ge_1), major_ge_3 ))]
+        const GDAL_OF_GNM = 0x08;
+        /// Allow multidimensional raster drivers to be used.
+        #[cfg(all(major_ge_3,minor_ge_1))]
+        const GDAL_OF_MULTIDIM_RASTER = 0x10;
+        /// Emit error message in case of failed open.
+        const GDAL_OF_VERBOSE_ERROR = 0x40;
+        /// Open as internal dataset. Such dataset isn't
+        /// registered in the global list of opened dataset.
+        /// Cannot be used with GDAL_OF_SHARED.
+        const GDAL_OF_INTERNAL = 0x80;
+
+        /// Default strategy for cached blocks.
+        #[cfg(any( all(major_ge_2,minor_ge_1), major_ge_3 ))]
+        const GDAL_OF_DEFAULT_BLOCK_ACCESS = 0;
+
+        /// Array based strategy for cached blocks.
+        #[cfg(any( all(major_ge_2,minor_ge_1), major_ge_3 ))]
+        const GDAL_OF_ARRAY_BLOCK_ACCESS = 0x100;
+
+        /// Hashset based strategy for cached blocks.
+        #[cfg(any( all(major_ge_2,minor_ge_1), major_ge_3 ))]
+        const GDAL_OF_HASHSET_BLOCK_ACCESS = 0x200;
+    }
+}
+
+impl Default for GdalOpenFlags {
+    fn default() -> GdalOpenFlags {
+        GdalOpenFlags::GDAL_OF_READONLY
+    }
+}
+
+impl From<GDALAccess::Type> for GdalOpenFlags {
+    fn from(val: GDALAccess::Type) -> GdalOpenFlags {
+        if val == GDALAccess::GA_Update {
+            GdalOpenFlags::GDAL_OF_UPDATE
+        } else {
+            GdalOpenFlags::GDAL_OF_READONLY
+        }
+    }
+}
+
+// Open parameters
+#[derive(Debug, Default)]
+pub struct DatasetOptions<'a> {
+    pub open_flags: GdalOpenFlags,
+    pub allowed_drivers: Option<&'a [&'a str]>,
+    pub open_options: Option<&'a [&'a str]>,
+    pub sibling_files: Option<&'a [&'a str]>,
 }
 
 // GDAL Docs state: The returned dataset should only be accessed by one thread at a time.
@@ -46,25 +139,23 @@ impl Dataset {
         self.c_dataset
     }
 
+    /// Open a dataset at the given `path` with default
+    /// options.
     pub fn open(path: &Path) -> Result<Dataset> {
-        Self::open_ex(path, None, None, None, None)
+        Self::open_ex(path, DatasetOptions::default())
     }
 
-    pub fn open_ex(
-        path: &Path,
-        open_flags: Option<GDALAccess::Type>,
-        allowed_drivers: Option<&[&str]>, // TODO: use parameters
-        open_options: Option<&[&str]>,
-        sibling_files: Option<&[&str]>,
-    ) -> Result<Dataset> {
+    /// Open a dataset with extended options. See
+    /// [GDALOpenEx].
+    pub fn open_ex(path: &Path, options: DatasetOptions) -> Result<Dataset> {
         _register_drivers();
         let filename = path.to_string_lossy();
         let c_filename = CString::new(filename.as_ref())?;
-        let c_open_flags = open_flags.unwrap_or(GDALAccess::GA_ReadOnly); // This defaults to GdalAccess::GA_ReadOnly
+        let c_open_flags = options.open_flags.bits;
 
         // handle driver params:
         // we need to keep the CStrings and the pointers around
-        let c_allowed_drivers = allowed_drivers.map(|d| {
+        let c_allowed_drivers = options.allowed_drivers.map(|d| {
             d.iter()
                 .map(|&s| CString::new(s))
                 .collect::<std::result::Result<Vec<CString>, NulError>>()
@@ -77,7 +168,7 @@ impl Dataset {
         let mut c_drivers_ptrs = c_drivers_vec.iter().map(|s| s.as_ptr()).collect::<Vec<_>>();
         c_drivers_ptrs.push(ptr::null());
 
-        let c_drivers_ptr = if allowed_drivers.is_some() {
+        let c_drivers_ptr = if options.allowed_drivers.is_some() {
             c_drivers_ptrs.as_ptr()
         } else {
             ptr::null()
@@ -85,7 +176,7 @@ impl Dataset {
 
         // handle open options params:
         // we need to keep the CStrings and the pointers around
-        let c_open_options = open_options.map(|d| {
+        let c_open_options = options.open_options.map(|d| {
             d.iter()
                 .map(|&s| CString::new(s))
                 .collect::<std::result::Result<Vec<CString>, NulError>>()
@@ -101,7 +192,7 @@ impl Dataset {
             .collect::<Vec<_>>();
         c_open_options_ptrs.push(ptr::null());
 
-        let c_open_options_ptr = if open_options.is_some() {
+        let c_open_options_ptr = if options.open_options.is_some() {
             c_open_options_ptrs.as_ptr()
         } else {
             ptr::null()
@@ -109,7 +200,7 @@ impl Dataset {
 
         // handle sibling files params:
         // we need to keep the CStrings and the pointers around
-        let c_sibling_files = sibling_files.map(|d| {
+        let c_sibling_files = options.sibling_files.map(|d| {
             d.iter()
                 .map(|&s| CString::new(s))
                 .collect::<std::result::Result<Vec<CString>, NulError>>()
@@ -125,7 +216,7 @@ impl Dataset {
             .collect::<Vec<_>>();
         c_sibling_files_ptrs.push(ptr::null());
 
-        let c_sibling_files_ptr = if sibling_files.is_some() {
+        let c_sibling_files_ptr = if options.sibling_files.is_some() {
             c_sibling_files_ptrs.as_ptr()
         } else {
             ptr::null()
@@ -154,11 +245,13 @@ impl Dataset {
         Dataset { c_dataset }
     }
 
+    /// Fetch the projection definition string for this dataset.
     pub fn projection(&self) -> String {
         let rv = unsafe { gdal_sys::GDALGetProjectionRef(self.c_dataset) };
         _string(rv)
     }
 
+    /// Set the projection reference string for this dataset.
     pub fn set_projection(&self, projection: &str) -> Result<()> {
         let c_projection = CString::new(projection)?;
         unsafe { gdal_sys::GDALSetProjection(self.c_dataset, c_projection.as_ptr()) };
@@ -166,11 +259,13 @@ impl Dataset {
     }
 
     #[cfg(major_ge_3)]
+    /// Get the spatial reference system for this dataset.
     pub fn spatial_ref(&self) -> Result<SpatialRef> {
         unsafe { SpatialRef::from_c_obj(gdal_sys::GDALGetSpatialRef(self.c_dataset)) }
     }
 
     #[cfg(major_ge_3)]
+    /// Set the spatial reference system for this dataset.
     pub fn set_spatial_ref(&self, spatial_ref: &SpatialRef) -> Result<()> {
         let rv = unsafe { gdal_sys::GDALSetSpatialRef(self.c_dataset, spatial_ref.to_c_hsrs()) };
         if rv != CPLErr::CE_None {
@@ -198,6 +293,7 @@ impl Dataset {
         Ok(unsafe { Dataset::from_c_dataset(c_dataset) })
     }
 
+    /// Fetch the driver to which this dataset relates.
     pub fn driver(&self) -> Driver {
         unsafe {
             let c_driver = gdal_sys::GDALGetDatasetDriver(self.c_dataset);
@@ -205,6 +301,10 @@ impl Dataset {
         }
     }
 
+    /// Fetch a band object for a dataset.
+    ///
+    /// Applies to raster datasets, and fetches the
+    /// rasterband at the given _1-based_ index.
     pub fn rasterband(&self, band_index: isize) -> Result<RasterBand> {
         unsafe {
             let c_band = gdal_sys::GDALGetRasterBand(self.c_dataset, band_index as c_int);
@@ -219,10 +319,15 @@ impl Dataset {
         unsafe { Layer::from_c_layer(self, c_layer) }
     }
 
+    /// Get the number of layers in this dataset.
     pub fn layer_count(&self) -> isize {
         (unsafe { gdal_sys::OGR_DS_GetLayerCount(self.c_dataset) }) as isize
     }
 
+    /// Fetch a layer by index.
+    ///
+    /// Applies to vector datasets, and fetches by the given
+    /// _0-based_ index.
     pub fn layer(&mut self, idx: isize) -> Result<Layer> {
         let c_layer = unsafe { gdal_sys::OGR_DS_GetLayer(self.c_dataset, idx as c_int) };
         if c_layer.is_null() {
@@ -231,6 +336,7 @@ impl Dataset {
         Ok(self.child_layer(c_layer))
     }
 
+    /// Fetch a layer by name.
     pub fn layer_by_name(&mut self, name: &str) -> Result<Layer> {
         let c_name = CString::new(name)?;
         let c_layer = unsafe { gdal_sys::OGR_DS_GetLayerByName(self.c_dataset(), c_name.as_ptr()) };
@@ -240,14 +346,17 @@ impl Dataset {
         Ok(self.child_layer(c_layer))
     }
 
+    /// Returns an iterator over the layers of the dataset.
     pub fn layers(&self) -> LayerIterator {
         LayerIterator::with_dataset(self)
     }
 
+    /// Fetch the number of raster bands on this dataset.
     pub fn raster_count(&self) -> isize {
         (unsafe { gdal_sys::GDALGetRasterCount(self.c_dataset) }) as isize
     }
 
+    /// Returns the raster dimensions: (width, height).
     pub fn raster_size(&self) -> (usize, usize) {
         let size_x = unsafe { gdal_sys::GDALGetRasterXSize(self.c_dataset) } as usize;
         let size_y = unsafe { gdal_sys::GDALGetRasterYSize(self.c_dataset) } as usize;
@@ -576,10 +685,11 @@ mod tests {
 
         let ds = Dataset::open_ex(
             &temp_path,
-            Some(GDALAccess::GA_Update),
-            Some(&["GPKG"]),
-            None,
-            None,
+            DatasetOptions {
+                open_flags: GDALAccess::GA_Update.into(),
+                allowed_drivers: Some(&["GPKG"]),
+                ..DatasetOptions::default()
+            },
         )
         .unwrap();
         (temp_path, ds)
@@ -598,10 +708,10 @@ mod tests {
     fn test_open_ex_ro_vector() {
         Dataset::open_ex(
             fixture!("roads.geojson"),
-            Some(GDALAccess::GA_ReadOnly),
-            None,
-            None,
-            None,
+            DatasetOptions {
+                open_flags: GDALAccess::GA_ReadOnly.into(),
+                ..DatasetOptions::default()
+            },
         )
         .unwrap();
     }
@@ -610,10 +720,10 @@ mod tests {
     fn test_open_ex_update_vector() {
         Dataset::open_ex(
             fixture!("roads.geojson"),
-            Some(GDALAccess::GA_Update),
-            None,
-            None,
-            None,
+            DatasetOptions {
+                open_flags: GDALAccess::GA_Update.into(),
+                ..DatasetOptions::default()
+            },
         )
         .unwrap();
     }
@@ -622,29 +732,60 @@ mod tests {
     fn test_open_ex_allowed_driver_vector() {
         Dataset::open_ex(
             fixture!("roads.geojson"),
-            None,
-            Some(&["GeoJSON"]),
-            None,
-            None,
+            DatasetOptions {
+                allowed_drivers: Some(&["GeoJSON"]),
+                ..DatasetOptions::default()
+            },
         )
         .unwrap();
     }
 
     #[test]
     fn test_open_ex_allowed_driver_vector_fail() {
-        Dataset::open_ex(fixture!("roads.geojson"), None, Some(&["TIFF"]), None, None).unwrap_err();
+        Dataset::open_ex(
+            fixture!("roads.geojson"),
+            DatasetOptions {
+                allowed_drivers: Some(&["TIFF"]),
+                ..DatasetOptions::default()
+            },
+        )
+        .unwrap_err();
     }
 
     #[test]
     fn test_open_ex_open_option() {
         Dataset::open_ex(
             fixture!("roads.geojson"),
-            None,
-            None,
-            Some(&["FLATTEN_NESTED_ATTRIBUTES=YES"]),
-            None,
+            DatasetOptions {
+                open_options: Some(&["FLATTEN_NESTED_ATTRIBUTES=YES"]),
+                ..DatasetOptions::default()
+            },
         )
         .unwrap();
+    }
+
+    #[test]
+    fn test_open_ex_extended_flags_vector() {
+        Dataset::open_ex(
+            fixture!("roads.geojson"),
+            DatasetOptions {
+                open_flags: GdalOpenFlags::GDAL_OF_UPDATE | GdalOpenFlags::GDAL_OF_VECTOR,
+                ..DatasetOptions::default()
+            },
+        )
+        .unwrap();
+    }
+
+    #[test]
+    fn test_open_ex_extended_flags_vector_fail() {
+        Dataset::open_ex(
+            fixture!("roads.geojson"),
+            DatasetOptions {
+                open_flags: GdalOpenFlags::GDAL_OF_UPDATE | GdalOpenFlags::GDAL_OF_RASTER,
+                ..DatasetOptions::default()
+            },
+        )
+        .unwrap_err();
     }
 
     #[test]