+++ src/lib.rs
@@ -11,7 +11,7 @@
 //! let mut dataset = Dataset::open(Path::new("fixtures/roads.geojson")).unwrap();
 //! let layer = dataset.layer(0).unwrap();
 //! for feature in layer.features() {
-//!     let highway_field = feature.field("highway").unwrap();
+//!     let highway_field = feature.field("highway").unwrap().unwrap();
 //!     let geometry = feature.geometry();
 //!     println!("{} {}", highway_field.into_string().unwrap(), geometry.wkt().unwrap());
 //! }
@@ -34,7 +34,9 @@ mod utils;
 pub mod vector;
 pub mod version;
 
-pub use dataset::{Dataset, GeoTransform, LayerIterator, Transaction};
+pub use dataset::{
+    Dataset, DatasetOptions, GdalOpenFlags, GeoTransform, LayerIterator, Transaction,
+};
 pub use driver::Driver;
 pub use metadata::Metadata;