+++ src/raster/tests.rs
@@ -443,6 +443,42 @@ fn test_get_rasterband_actual_block_size() {
 }
 
 #[test]
+fn test_read_overviews() {
+    let dataset = Dataset::open(fixture!("tinymarble.tif")).unwrap();
+    let rasterband = dataset.rasterband(1).unwrap();
+    let overview_count = rasterband.overview_count().unwrap();
+    assert_eq!(overview_count, 2);
+
+    let overview_2 = rasterband.overview(0).unwrap();
+    let overview_4 = rasterband.overview(1).unwrap();
+    assert_eq!(rasterband.size(), (100, 50));
+
+    assert_eq!(overview_2.size(), (50, 25));
+    assert_eq!(overview_4.size(), (25, 13));
+}
+
+#[test]
+fn test_fail_read_overviews() {
+    let dataset = Dataset::open(fixture!("offset_scaled_tinymarble.tif")).unwrap();
+    let rasterband = dataset.rasterband(1).unwrap();
+    let overview_count = rasterband.overview_count().unwrap();
+    assert_eq!(overview_count, 0);
+
+    let overview_2 = rasterband.overview(0);
+    assert!(overview_2.is_err());
+}
+
+#[test]
+fn test_rasterband_lifetime() {
+    let dataset: Dataset = Dataset::open(fixture!("tinymarble.tif")).unwrap();
+    let rasterband = dataset.rasterband(1).unwrap();
+    let overview = rasterband.overview(0).unwrap();
+
+    drop(rasterband);
+    assert!(overview.no_data_value().is_none());
+}
+
+#[test]
 fn test_get_rasterband_color_interp() {
     let dataset = Dataset::open(fixture!("tinymarble.png")).unwrap();
     let rasterband = dataset.rasterband(1).unwrap();