+++ src/vector/mod.rs
@@ -9,7 +9,7 @@
 //! let mut dataset = Dataset::open(Path::new("fixtures/roads.geojson")).unwrap();
 //! let layer = dataset.layer(0).unwrap();
 //! for feature in layer.features() {
-//!     let highway_field = feature.field("highway").unwrap();
+//!     let highway_field = feature.field("highway").unwrap().unwrap();
 //!     let geometry = feature.geometry();
 //!     println!("{} {}", highway_field.into_string().unwrap(), geometry.wkt().unwrap());
 //! }